+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Fixed a deadlock on update for multiple clients targeting the same local file.
+
 ## [1.11.0] - 2024-02-01
 
 ### Added